Unit Test: reset mock function before each test case
The mock function will record each result that have be happend.
The function toHaveBeenCalled() will seach the result from the all
records.
This is make sense, because sometime the mock will be call mulitipl
times with one time. And we expect to test then.
So we need to reset the mockFn each time.

And I don't want the unit test print too many log.